+++ core/state/contracts/poly.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: LGPL-3.0
 pragma solidity >=0.5.0;
 
 // solc --allow-paths ., --abi --bin --overwrite --optimize -o core/state/contracts/build core/state/contracts/poly.sol
